saving records in the product_order table

+++ src/main/java/com/management/pizzaria/dtos/ProductOrderDTO.java
@@ -6,8 +6,8 @@ import com.management.pizzaria.models.Product;
 import java.util.List;
 
 public record ProductOrderDTO(
-        int orderId,
-        int productId,
+        Long orderId,
+        Long productId,
         int quantity,
         List<ProductOrderDTO> itemsOrder
 ) {
